requirements.txt
python-decouple==3.8
discord==2.3.2
pydo
